game-pipeline/requirements.txt
requests==2.25.1
psycopg2==2.8.6
kafka-python==2.0.2